Develop

+++ src/main/java/com/project/awesomegroup/repository/TeamRepository.java
@@ -1,10 +1,10 @@
 package com.project.awesomegroup.repository;
 
-import com.project.awesomegroup.dto.Team;
+import com.project.awesomegroup.dto.team.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 
 }
